refactor: remove BlockFetcher module and integrate its functionality into EvmRpcSource
