+++ src/providers/mod.rs
@@ -1,5 +1,4 @@
 //! Provider modules for various data sources.
 
-pub mod block_fetcher;
 pub mod rpc;
 pub mod traits;
